+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Allow a custom texture for each of the color picker colors.
 * Allow a custom palette for each of the color picker colors.
+* Added support for prismatic coloring.
+
 
 ### Changed